Refactor import resource script
